+++ internal/provider/cluster_data_source.go
@@ -157,6 +157,10 @@ func (d *clusterDataSource) Schema(
 							Computed:    true,
 							Description: "Internal DNS name of the cluster within the cloud provider's network. Used to connect to the cluster with PrivateLink or VPC peering.",
 						},
+						"private_endpoint_dns": schema.StringAttribute{
+							Computed:    true,
+							Description: "Domain name of the cluster for the private endpoint connection. This DNS name is used by GCP Private Service Connect to connect to the cluster.",
+						},
 						"node_count": schema.Int64Attribute{
 							Computed:    true,
 							Description: "Number of nodes in the region. Will always be 0 for serverless clusters.",